Create README.md
The project presents the circuit synthesis of a MIPS (Microprocessor without Interlocked Pipelined Stages) in a single-cycle format, involving the execution of all processor stages (instruction fetch, decode, execute, memory access, and write-back) for each individual instruction before moving on to the next one. This is a straightforward and easily implementable approach but lacks the advantages of parallel processing, which can impact performance with complex programs. In addition to the classical MIPS processor implementation, a program for weighted arithmetic mean has been added.